requirements.txt
groq
pandas
python-dotenv
sentence-transformers
scikit-learn
rapidfuzz
nltk
nbconvert
